oaktree: Show changed files, delete old files.
OakTree now deletes old files when generating its content. This only
applies to HTML files under public/, so it should be safe enough.
(Of course this will later prove to be unsafe.)

In addition, the output now shows which files are new, which have been
updated, and which will be deleted. In this case, a + indicates a new
file, a * indicates an updated file, and a - indicates a deleted file.
Skipped files are no longer displayed.