Add date_path_format spec option.
This controls how the date is formatted in post paths, so you can now
change that to whatever you want (currently set up to be similar to
WordPress). You do not have to end this with a forward slash, but it
should never start with one (the date always follows another set
path). So, if you wanted to, you could have all your posts in flat
files rather than indexed by year and month.