Add kind field to post heads, add static type.
Static posts are essentially posts that exist outside the usual
chronological layout of a blog -- they're still there, and still sorted
by their date (TODO: maybe add another way to sort them), but otherwise
they're just regular posts.

They end up in the <base>/<slug>/ directory without date directory
formatting and all that. These are sort of analogous to WordPress pages
in that they aren't part of the regular flow of posts and would likely
be used for static informational content that's secondary to the blog
itself.

They're kind of a hack on top of the existing posts right now, so
that's something to keep in mind.